requirements.txt
alembic==1.7.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-generator==1.10
attrs==21.4.0
Automat==0.6.0
azure-core==1.24.2
azure-storage-blob==12.13.0
backcall==0.2.0
bleach==4.1.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloud-init==20.2
cloudpickle==2.0.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.11.0
databricks-cli==0.16.6
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro-info===0.18ubuntu0.18.04.1
docker==5.0.3
e==1.4.5
entrypoints==0.4
example-package==0.0.1
example-package-ShankarSreenu @ file:///mnt/c/Users/vanam.sreenu/Downloads/Intern/build%20and%20deploy/package_tutorial
filelock==3.4.1
Flask==2.0.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.20
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==16.10.1
invoke==1.7.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
luminol==0.4
Mako==1.1.6
mando==0.6.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mle-training @ file:///mnt/c/Users/vanam.sreenu/Downloads/Intern/mle_training
mlflow==1.23.1
msrest==0.7.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.1
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
Pygments==2.12.0
PyGObject==3.26.1
PyJWT==2.3.0
pyOpenSSL==17.5.0
pyparsing==3.0.8
pyrsistent==0.18.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2022.1
pyxdg==0.25
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.2.2
QtPy==2.0.1
querystring-parser==1.2.4
radon==5.1.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
ruptures==1.1.3
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.8.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
ssh-import-id==5.7
systemd-python==234
ta-lib @ file:///mnt/c/Users/vanam.sreenu/Downloads/Intern/final_module/regression-py/regression-py
tabulate==0.8.9
terminado==0.13.0
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==4.3.3
Twisted==17.9.0
typing_extensions==4.1.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webdriver-manager==3.4.1
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
XlsxWriter==1.3.8
zipp==3.6.0
zope.interface==4.3.2
